% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap.data.frame}
\alias{bootstrap.data.frame}
\title{Bootstrap constructor}
\usage{
\method{bootstrap}{data.frame}(data, FUN = mean, n = 1000, ...)
}
\arguments{
\item{data}{The data frame from which to resample}

\item{FUN}{The function of interest}

\item{n}{The number of resamples to perform}

\item{...}{Additional arguments}
}
\value{
A an object with class \code{strop}
}
\description{
Bootstrap constructor
}
\examples{
data(mtcars)
bootstrap(mtcars, function(df) mean(df$cyl))
}
