% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.strop}
\alias{confint.strop}
\title{Symmetric bootstrap confidence intervals}
\usage{
\method{confint}{strop}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{A bootstrap object}

\item{parm}{Ignored}

\item{level}{The confidence level for the interval}

\item{...}{Additional arguments passed to \code{quantile}}
}
\value{
A symmetric \code{1 - level} confidence interval based on bootstrap resampling
}
\description{
Symmetric bootstrap confidence intervals
}
\examples{
data(mtcars)
obj <- bootstrap(mtcars$cyl)
confint(obj)
}
