% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrboot.R
\name{mlrboot}
\alias{mlrboot}
\title{Multiple linear regression bootstrapping}
\usage{
mlrboot(fit, ...)
}
\arguments{
\item{fit}{A fitted model (e.g. \code{lm}, \code{lm}) object}

\item{...}{Additional parameters}
}
\value{
An object of class \code{strop}
}
\description{
Multiple linear regression bootstrapping
}
\examples{
data(mtcars)
model <- lm(cyl ~ ., mtcars)
mlrboot(model)
}
