% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.strop}
\alias{plot.strop}
\title{Bootstrap visualization}
\usage{
\method{plot}{strop}(x, conf.level = 0.95, combine = F, vars = NA,
  ...)
}
\arguments{
\item{x}{A boostrap object of class \code{strop}}

\item{conf.level}{The confidence level for which to indicate central confidence intervals on each histogram}

\item{combine}{Whether to attempt to combine plots}

\item{vars}{The variables for which to produce plots}

\item{...}{Additional arguments}
}
\value{
\code{invisible(x)}
}
\description{
Bootstrap visualization
}
\examples{
data(mtcars)
plot(bootstrap(mtcars$cyl))
}
