% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.strop}
\alias{print.strop}
\title{Bootstrap result printing}
\usage{
\method{print}{strop}(x, max = 10, ...)
}
\arguments{
\item{x}{A boostrap object of class \code{strop}}

\item{max}{The maximum number of bootstrap estimates to print}

\item{...}{Additional arguments}
}
\description{
Bootstrap result printing
}
\examples{
data(mtcars)
print(bootstrap(mtcars$cyl))
}
