% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile.strop}
\alias{quantile.strop}
\title{Bootstrap quantiles}
\usage{
\method{quantile}{strop}(x, probs, ...)
}
\arguments{
\item{x}{A bootstrap object}

\item{probs}{Vector of probabilities}

\item{...}{Additional arguments passed to \code{quantile}}
}
\value{
A vector of quantiles
}
\description{
Bootstrap quantiles
}
\examples{
data(mtcars)
obj <- bootstrap(mtcars$cyl)
quantile(obj, 0.5)
}
