% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap constructor}
\usage{
bootstrap(data, FUN = mean, ...)
}
\arguments{
\item{data}{The data from which to resample}

\item{FUN}{The function of interest}

\item{...}{Additional arguments}
}
\description{
Bootstrap constructor
}
